Changed BrightSensor to transmit a 16-Bit value

+++ EclipseWS/SensrChip/BrightSensor.h
@@ -16,7 +16,7 @@
 
 class BrightSensor : TWI_S_Job {
 private:
-	uint8_t bLevel = 0;
+	uint16_t bLevel = 0;
 
 public:
 	BrightSensor();
